requirements.txt
streamlit==1.45.1
pandas==2.2.2
scikit-learn==1.4.0
joblib==1.3.2
pytest==8.3.5
pytest-cov==6.1.1
numpy==1.26.4
flake8==7.0.0
pyyaml==6.0.1
